# perf: compute inverse of 2 only once in sqrt