Yet another repeated inverse removal.

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added/run

Benchmark results:
```
Stark FP operations/sqrt                                                                             
                        time:   [5.3247 µs 5.3269 µs 5.3293 µs]
                        change: [+2.6661% +2.7406% +2.8178%] (p = 0.00 < 0.05)
                        Performance has regressed.

Stark FP operations/sqrt squared                                                                            
                        time:   [220.20 µs 220.26 µs 220.33 µs]
                        change: [-38.188% -38.152% -38.117%] (p = 0.00 < 0.05)
                        Performance has improved.
```